Add new SpatialLabs3D subproject for DLL.
This will create a DLL that we can LoadLibrary on, and thus remove our need to statically link against all the SR libs.

The files in SR have been moved to this subproject folder, and added to the subproject.

Does not currently compile, this is just the basic structure and where the files need to be.