Setup SpatialLabs3D.dll export functions.
The SpatialLabs3D unit will be the LoadLibrary/GetProcAddress piece for handling dynamic loading of the dll.  The output functions are declared "__declspec(dllexport) void __cdecl" so that we can easily GetProcAddress on them later.

This subproject now correctly generates a SpatialLabs3D.dll with the exported routines.  So it will be a standalone item loaded dynamically via LoadLibrary, and thus no longer causing d3d11.dll to have load time dependencies on the SR and Dimenco dlls.