Setup HackerDXGI to use SpatialLabs3D.dll.
Restructure the HackerSwapChainDirectMode::Init so that after it sets up the overlay, we look for gDirectMode == SpatialLabs and if so, we LoadLibrary on SpatialLabs3D.dll, and find the functions with GetProcAddress for StartSRWeaver, GetRTV, and Render.

Assuming those work out to be valid, we can then use them during FrameActions to route the frame data to the SR render libraries.

So this successfully removes the load time requirement of having all the Dimenco and SR dlls available, and is only dependent on the existence of our SpatialLabs3D.dll.

It also now works under x32 without #ifdef, because the library will always be missing under x32. If the direct_mode=spatial_labs in this case, it will report an error in the overlay.