Switch from __cdecl to default __stdcall.
Not sure this matters, but from reading StackOverflow it seems like we do not want to use __cdecl for these, and it might set us up for stack misalignment issues.  __stdcall is the default for Win32 programming, so let's just allow it use default.

On x64 I think this doesn't matter, because the calling conventions are ignored in favor of register based parameters. And this is x64 only, so probably this will not change either way.