Set up proper build outputs for SpatialLabs3D.dll
This needs to be built into the same output folder as the rest of the components.  x64/Release or x64/Debug, and is skipped for x32 builds.

Set tool chain to match the DirectX11 project. No need to use a variety of tools and add complexity.

Delete the xcopy of SR library dlls to the output folder, as now they are no longer needed at runtime.

Remove the SpatialLabs3D.dll  x32 builds from Configurations, because the SR libraries are x64 only.

This now builds all 6 targets successfully.