Update project files for new SDK
Changes all the project include paths for the added x32 or x64 path in the SDK. For this pass, just fixing x64 build.

Needs a change to the SRWeaver creation- they deprecated all the prior constructors and need HWND now. Added that to the SR Info struct at creation.

Compiles but is untested.