Convert dll to SR.lib output
This will now generate an SR.lib output file for use in simulated_reality output mode. Instead of being a dll that gets carted around, this will be a header file and static library to be linked against.

Limited testing, but it works in AC4 both with and without simulated reality dlls.