Rename x86 target to x32.
Keeping with the philosophy of trying to make it more clear, we are using x32/x64 as the bitwise builds.

Win32 is just stupid, and belong nowhere. x86 is archaic and nonsensical.

Cannot fix the actual configuration parameter of Win32, hardcoded in VS, but anything user/programmer facing should be x32.

This fixes the build targets.