Update to VS2022, toolset v143, on windows-2025
We want to use the Github runners, and they are forcing people to update. Since we have to change toolsets, lets update to the latest available.

Siwtching to windows-2025 as latest available.

We will still specify exact toolsets, not 'latest' because we don't want tools changing out from underneath us.